build - Adjust _CMSG_ALIGN() for compatibility
* The argument to _CMSG_ALIGN() is commonly unsigned, fix compiler
  warnings by making the alignment constants unsigned.

* Also take a page from FreeBSD and cast the (n) argument to an
  unsigned quantity (size_t).

* Note that the alignment code is still hardwired to 4 bytes,
  and will eventually have to be changed to something more portable.
  For now do not change it as doing so would introduce an ABI
  incompatibility.

Rerported-by: rsmarples (Roy Marples)

+++ sys/sys/socket.h
@@ -425,7 +425,7 @@ struct cmsgcred {
 };
 
 /* Alignment requirement for CMSG struct manipulation */
-#define	_CMSG_ALIGN(n)		(((n) + 3) & ~3)
+#define	_CMSG_ALIGN(n)		(((size_t)(n) + 3U) & ~3U)
 
 #ifdef _KERNEL
 #define	CMSG_ALIGN(n)		_CMSG_ALIGN(n)
